fibonacci 50

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,12 +7,12 @@
 int main(void)
 {
 	int i;
-	long long int fn_1 = 1, fn_2 = 0, f;
+	long int fn_1 = 1, fn_2 = 0, f;
 
 	for (i = 2; i <= 50; i++)
 	{
 		f = fn_2 + fn_1;
-		printf("%lld", f);
+		printf("%ld", f);
 		fn_2 = fn_1;
 		fn_1 = f;
 		if (i < 50)
